Move unsupported Heal related functions to azure-unsupported.go
